+++ src/test/java/lotto/number/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.number;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
